pyproject.toml
[project]
name = "invoicia-notif"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Msa360", email = "arnaud25@icloud.com" }
]
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
